Remove useless if
Before, this thing was first checking if the matche's state is FINISHED and then if it checked if the match is not RUNNING, which only could mean the match hasn't begun yet. Since in both cases, literally nothing is done, I think it would be better to remove the first if clause and update the message in the second.